requirements.txt
discord.py==1.7.3
SQLAlchemy
python-dotenv
docker
psycopg2
transformers
nltk
openai
python-qbittorrent
arrow
promise
redis
innertube
pytube